refactored a little bit and handles existing ingredients in csv
